requirements.txt
numpy>=1.19.2
PyQt5>=5.15.0
pyqtgraph>=0.12.0
tqdm>=4.50.0
pyserial>=3.5
scipy>=1.6.0 